Week4_FastAPI/hw4_2/requirements.txt
annotated-types      0.7.0
anyio                4.8.0
bcrypt               4.2.1
certifi              2025.1.31
click                8.1.8
dnspython            2.7.0
email_validator      2.2.0
fastapi              0.115.5
fastapi-cli          0.0.7
h11                  0.14.0
httpcore             1.0.7
httptools            0.6.4
httpx                0.28.1
idna                 3.10
itsdangerous         2.2.0
Jinja2               3.1.5
markdown-it-py       3.0.0
MarkupSafe           3.0.2
mdurl                0.1.2
orjson               3.10.15
passlib              1.7.4
pip                  25.0
psycopg2-binary      2.9.10
pydantic             2.10.6
pydantic_core        2.27.2
pydantic-extra-types 2.10.2
pydantic-settings    2.7.1
Pygments             2.19.1
python-dotenv        1.0.1
python-multipart     0.0.20
PyYAML               6.0.2
rich                 13.9.4
rich-toolkit         0.13.2
setuptools           75.8.0
shellingham          1.5.4
sniffio              1.3.1
SQLAlchemy           2.0.38
sqlmodel             0.0.22
starlette            0.41.3
typer                0.15.1
typing_extensions    4.12.2
ujson                5.10.0
uvicorn              0.32.0
uvloop               0.21.0
watchfiles           1.0.4
websockets           14.2
wheel                0.45.1